unified tabs and quotes

+++ ztlearn/utils/im2col.py
@@ -36,11 +36,11 @@ def get_im2col_indices(x_shape, field_height = 3, field_width = 3, padding = ((0
     out_height = (H + np.sum(pad_height) - field_height) / stride + 1
     out_width = (W + np.sum(pad_width) - field_width) / stride + 1
 
-    i0 = np.repeat(np.arange(field_height, dtype='int32'), field_width)
+    i0 = np.repeat(np.arange(field_height, dtype = 'int32'), field_width)
     i0 = np.tile(i0, C)
-    i1 = stride * np.repeat(np.arange(out_height,dtype='int32'), out_width)
+    i1 = stride * np.repeat(np.arange(out_height, dtype = 'int32'), out_width)
     j0 = np.tile(np.arange(field_width), field_height * C)
-    j1 = stride * np.tile(np.arange(out_width, dtype='int32'), int(out_height))
+    j1 = stride * np.tile(np.arange(out_width, dtype = 'int32'), int(out_height))
     i = i0.reshape(-1, 1) + i1.reshape(1, -1)
     j = j0.reshape(-1, 1) + j1.reshape(1, -1)
 
@@ -48,10 +48,10 @@ def get_im2col_indices(x_shape, field_height = 3, field_width = 3, padding = ((0
 
     return (k, i, j)
 
-def im2col_indices(x, field_height, field_width, padding, stride=1):
+def im2col_indices(x, field_height, field_width, padding, stride = 1):
     """ An implementation of im2col based on some fancy indexing """
     pad_height, pad_width = padding
-    x_padded = np.pad(x, ((0, 0), (0, 0), pad_height, pad_width), mode='constant')
+    x_padded = np.pad(x, ((0, 0), (0, 0), pad_height, pad_width), mode = 'constant')
     k, i, j = get_im2col_indices(x.shape, field_height, field_width, padding, stride)
     cols = x_padded[:, k, i, j]
     C = x.shape[1]
@@ -59,12 +59,12 @@ def im2col_indices(x, field_height, field_width, padding, stride=1):
 
     return cols
 
-def col2im_indices(cols, x_shape, field_height=3, field_width=3, padding = ((0, 0), (0, 0)), stride=1):
+def col2im_indices(cols, x_shape, field_height = 3, field_width = 3, padding = ((0, 0), (0, 0)), stride=1):
     """ An implementation of col2im based on fancy indexing and np.add.at """
     N, C, H, W = x_shape
     pad_height, pad_width = padding
     H_padded, W_padded = H + np.sum(pad_height), W + np.sum(pad_width)
-    x_padded = np.zeros((N, C, H_padded, W_padded), dtype=cols.dtype)
+    x_padded = np.zeros((N, C, H_padded, W_padded), dtype = cols.dtype)
     k, i, j = get_im2col_indices(x_shape, field_height, field_width, padding, stride)
     cols_reshaped = cols.reshape(C * field_height * field_width, -1, N)
     cols_reshaped = cols_reshaped.transpose(2, 0, 1)
